--- post_truth/.google-cookie
@@ -1,1 +1,0 @@
-#LWP-Cookies-2.0
